clean up and rename files
