+++ nuc_morph_analysis/analyses/density/run_all_density_workflows.py
@@ -1,6 +1,9 @@
-from nuc_morph_analysis.analyses.density import watershed_validate
-watershed_validate.run_validation_and_plot(plot_everything=False)
-from nuc_morph_analysis.analyses.density import effect_of_resolution_level
-from nuc_morph_analysis.analyses.density import pseudocell_based_density_analysis
+# this code runs all workflows in the density folder (not all of these are in the paper)
+from nuc_morph_analysis.analyses.density import visually_validate_watershed_psuedo_cell_seg_workflow
+visually_validate_watershed_psuedo_cell_seg_workflow.run_validation_and_plot(plot_everything=False)
+from nuc_morph_analysis.analyses.density import confirm_identical_density_at_different_zarr_resolutions
+from nuc_morph_analysis.analyses.density import figure_watershed_based_density_schematic
+figure_watershed_based_density_schematic.run_validation_and_plot()
+from nuc_morph_analysis.analyses.density import confirm_no_corr_btn_density_and_nuclear_area
 
 print('done')
